Refactor shared library/module versioning logic to SCons/Tool/linkCommon/
